chore: refactor recipe editor
